日志库

+++ logger/logger_test.go
@@ -70,8 +70,8 @@ func TestCustomPrint(t *testing.T) {
 	var _addr addr
 	PrintlnWithAddr(Kick|Ping|Cli|Msg, _addr, "testing for 3")
 	Println(Kick|Ping|Cli|Addr|Msg, "testing for 3")
-	PrintfWithAddr(Kick|Ping|Cli|Msg, _addr, "testing for %d", 3)
-	Printf(Kick|Ping|Cli|Addr|Msg, "testing for %d", 3)
+	PrintfWithAddr(Kick|Ping|Cli|Msg, _addr, "testing for %d", uint64(3))
+	Printf(Kick|Ping|Cli|Addr|Msg, "testing for %d", uint64(3))
 	PrintlnWithAddr(Kick|Ping|Cli|Addr|Msg, _addr, "testing")
 	fmt.Printf("%b\n", Addr)
 	fmt.Printf("%x\n", Addr)
